+++ crates/zng-layout/src/unit/resolution.rs
@@ -34,20 +34,40 @@ impl std::hash::Hash for Ppi {
         super::about_eq_hash(self.0, 0.0001, state)
     }
 }
+impl Ord for Ppi {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        super::about_eq_ord(self.0, other.0, 0.0001)
+    }
+}
+impl PartialOrd for Ppi {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
 
 /// Pixels-per-meter resolution.
-#[derive(Debug, Clone, Copy, PartialEq, serde::Serialize, serde::Deserialize, Transitionable)]
+#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize, Transitionable)]
 #[serde(transparent)]
 pub struct Ppm(pub f32);
-impl Ppm {
-    /// Returns the minimum of the two resolutions.
-    pub fn min(self, other: impl Into<Ppm>) -> Ppm {
-        Ppm(self.0.min(other.into().0))
+impl PartialEq for Ppm {
+    fn eq(&self, other: &Self) -> bool {
+        super::about_eq(self.0, other.0, 0.0001)
     }
-
-    /// Returns the maximum of the two resolutions.
-    pub fn max(self, other: impl Into<Ppm>) -> Ppm {
-        Ppm(self.0.max(other.into().0))
+}
+impl Eq for Ppm {}
+impl std::hash::Hash for Ppm {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        super::about_eq_hash(self.0, 0.0001, state)
+    }
+}
+impl Ord for Ppm {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        super::about_eq_ord(self.0, other.0, 0.0001)
+    }
+}
+impl PartialOrd for Ppm {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
     }
 }
 impl Default for Ppm {